login and sign up pages to its own page, isolated from components folder
